Breakout database functionality into its own module
