+++ mmeds/authentication.py
@@ -1,6 +1,6 @@
 import hashlib
 from string import ascii_uppercase, ascii_lowercase
-from mmeds.database import Database
+from mmeds.database.database import Database
 from mmeds.config import STORAGE_DIR, get_salt
 from mmeds.util import send_email, log
 from mmeds.error import NoResultError, InvalidLoginError, InvalidPasswordErrors, InvalidUsernameError
